movie_app/requirements.txt
Django==4.2.16
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.5.0
django-redis==5.4.0
django-tinymce==4.1.0
sqlparse==0.5.1
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
